refactor(fuse): isolate recluster orchestration
